fsadfads

+++ rotasi.h
@@ -1,5 +1,5 @@
 #include "point_warna.h"
 
-point rotasi(point p1, point p2, int degree);
+point rotasi(point p1, point p2, double degree);
 
-void rotasiBanyak(point p0, point* p1, int degree, int banyak);
+void rotasiBanyak(point p0, point* p1, int degree, double banyak);
